+++ features/inventory/showArmorAttributes.js
@@ -9,7 +9,7 @@ register('renderItemIntoGui', (item, x, y, event) => {
 
 const FISHING_GEAR_REGEX = /(Thunder|Thunderbolt|Magma Lord|Slug|Moogma|Flaming|Taurus) (Helmet|Chestplate|Leggings|Boots|Gauntlet|Necklace)/;
 const CRIMSON_ARMOR_REGEX = /(Crimson|Aurora|Terror|Fervor|Hollow|Berserker|Rampart) (Helmet|Chestplate|Leggings|Boots)/;
-const CRIMSON_EQUIPMENT_REGEX = /Gauntlet of Contagion|Flaming Fist|((Molten|Implosion|Blaze|Scoville|Scourge|Delirium|Ghast|Lava Shell|Magma|Glowstone) (Cloak|Belt|Necklace|Gauntlet|Bracelet))/;
+const CRIMSON_EQUIPMENT_REGEX = /Gauntlet of Contagion|Flaming Fist|((Molten|Implosion|Blaze|Scoville|Scourge|Delirium|Ghast|Lava Shell|Tera Shell|Magma|Glowstone) (Cloak|Belt|Necklace|Gauntlet|Bracelet))/;
 
 function showArmorAttributes(item, x, y) {
     if (!item || (!settings.showFishingArmorAttributes && !settings.showCrimsonArmorAttributes) || !isInSkyblock()) {