+++ constants/fishingProfitItems.js
@@ -1568,6 +1568,27 @@ export const FISHING_PROFIT_ITEMS = [
         npcPrice: 2500000,
     },
 
+    // The Abandoned Quarry
+
+    {
+        itemId: 'MITHRIL_ORE',
+        itemName: 'Mithril',
+        itemDisplayName: `${COMMON}Mithril`,
+        npcPrice: 8,
+    },
+    {
+        itemId: 'ENCHANTED_MITHRIL',
+        itemName: 'Enchanted Mithril',
+        itemDisplayName: `${RARE}Enchanted Mithril`,
+        npcPrice: 1280,
+    },
+    {
+        itemId: 'SPECKLED_TEACUP',
+        itemName: 'Speckled Teacup',
+        itemDisplayName: `${SPECIAL}Speckled Teacup`,
+        npcPrice: 0,
+    },
+
     // Runes
 
     {