+++ docs/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 Released: ???
 
+Features:
+
+- Add 4 variations of Mithril Grubbers to sea creatures count widget.
+- Added mithril, enchanted mithril and specked teacup to the fishing profit tracker.
+- Render attributes on Tera Shell Necklace if the setting is enabled (same way as for other equipment).
+
 ## v1.24.0
 
 Released: 2025-01-22